+++ CHANGELOG.markdown
@@ -3,6 +3,13 @@
 
 ## Version 0.9.3 - *Not Yet Released*
 
+ * [Aphid.Model] Added support for manually parsing the instance or response
+   collections in each model class definition by defining parseCollectionResponse
+   or parseInstanceResponse.
+
+ * [Aphid.UI.ModalView] Modal views may now be dismissed by pressing the
+   escape key.
+
  * [Tools] Added "aphid vendorize" for vendorizing Aphid in an existing Aphid
    project.